Changing to RMS-based compression
RMS-based compression is supposed to work a lot better
for human hearing than peak-based compression.

The goal of this project is to equalize volume, so it will be used.

Additionally, the up compression has been removed as it will be
unecessary. Upward compression will be handled by an equalization step.